MyMathCalc: Should return double, not int

+++ src/main/java/com/example/springdocker/MyMatchCalc.java
@@ -9,7 +9,7 @@ public class MyMatchCalc {
         return a*b;
     }
 
-    public float divide(int a, int b){
+    public double divide(int a, int b){
         return a/b;
     }
 }
